Factor out socket RPC code into SocketAdapter
This is the start of de-duplicating some of our net.Socket based RPC
adapoter code. It creates a new SocketAdapter and renames all the
underlying types from IPC to socket adapter based type names. It also
makes TcpAdapter and TlsAdapter both use it.
