+++ ironfish/src/fileStores/config.ts
@@ -267,8 +267,8 @@ export class Config extends KeyStore<ConfigOptions> {
       rpcTcpPort: 8020,
       rpcTcpSecure: false,
       rpcRetryConnect: false,
-      tlsKeyPath: files.resolve(files.join(dataDir, 'node-key.pem')),
-      tlsCertPath: files.resolve(files.join(dataDir, 'node-cert.pem')),
+      tlsKeyPath: files.resolve(files.join(dataDir, 'certs', 'node-key.pem')),
+      tlsCertPath: files.resolve(files.join(dataDir, 'certs', 'node-cert.pem')),
       maxPeers: 50,
       minimumBlockConfirmations: 12,
       minPeers: 1,
