+++ ironfish/src/fileSystems/fileSystem.ts
@@ -15,5 +15,6 @@ export abstract class FileSystem {
   abstract mkdir(path: string, options: { recursive?: boolean }): Promise<void>
   abstract resolve(path: string): string
   abstract join(...paths: string[]): string
+  abstract dirname(path: string): string
   abstract exists(path: string): Promise<boolean>
 }
