+++ ironfish/src/fileSystems/nodeFileSystem.ts
@@ -53,6 +53,11 @@ export class NodeFileProvider extends FileSystem {
     return this.path.join(...paths)
   }
 
+  dirname(path: string): string {
+    Assert.isNotNull(this.path, `Must call FileSystem.init()`)
+    return this.path.dirname(path)
+  }
+
   async exists(path: string): Promise<boolean> {
     Assert.isNotNull(this.path, `Must call FileSystem.init()`)
     return await this.access(path)
