+++ ironfish/src/rpc/adapters/index.ts
@@ -6,5 +6,5 @@ export * from './adapter'
 export * from './errors'
 export * from './ipcAdapter'
 export * from './tcpAdapter'
-export * from './secureTcpAdapter'
+export * from './tlsAdapter'
 export * from './memoryAdapter'
